feat(api): Add overloaded listFiles method to StorageOperations interface
Introduces a new `default` method, `listFiles(String bucketName)`, to the `StorageOperations` interface.

Previously, calling `listFiles` required users to manually provide an empty string for the `prefix` parameter to list all files in a bucket. This was verbose and not intuitive for a very common use case.

This new overloaded method acts as a convenient shortcut, automatically calling the two-parameter `listFiles(bucketName, "")` method. By adding it to the interface, this functionality becomes an official part of the public contract, available to all consumers and implementations, thereby improving the overall API ergonomics.